refactor: move catchup types and merge functions to ingest_catchup.go
Move BatchChanges struct, mergeTrustlineChanges, mergeAccountChanges,
mergeSACBalanceChanges, and processBatchChanges to dedicated catchup file.

Zero behavior change — pure file reorganization.
